requirements.txt
torch
torchaudio
numpy
PyYAML
scipy
soundfile
tqdm
tensorboard
